@sushiswap/router
- Added packages/router and extracted relevent TypeScript code.
- Updated protocols/route-processor to use packages/router in testing.
- Updated apis/swap to use packages/router.
- Removed duplicate code.

IMPORTANT NODE: because this code was originally kept inside of a hardhat project with less than ideal TypeScript config,I've had to initially add @ts-nocheck to most files.

We need to remove these and fix all type errors ASAP.
