+++ packages/router/.eslintrc.js
@@ -1,0 +1,1 @@
+module.exports = require('@sushiswap/eslint-config')
