+++ packages/router/.prettierrc.js
@@ -1,0 +1,3 @@
+const { default: config } = require("@sushiswap/prettier-config");
+
+module.exports = config;
