+++ packages/router/README.md
@@ -1,0 +1,3 @@
+# Router
+
+Router library
