+++ packages/router/package.json
@@ -1,0 +1,70 @@
+{
+  "name": "@sushiswap/router",
+  "version": "0.0.0",
+  "private": true,
+  "description": "Sushi Router",
+  "keywords": [
+    "sushi",
+    "routers"
+  ],
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/sushiswap/sushiswap.git",
+    "directory": "packages/router"
+  },
+  "license": "MIT",
+  "author": "Matthew Lilley <hello@matthewLilley.com>",
+  "main": "dist/index.js",
+  "source": "src/index.ts",
+  "typings": "dist/index.d.ts",
+  "exports": {
+    ".": {
+      "types": "./dist/index.d.ts",
+      "default": "./dist/index.js"
+    },
+    "./liquidity-providers/Trident": {
+      "types": "./dist/liquidity-providers/Trident.d.ts",
+      "default": "./dist/liquidity-providers/Trident.js"
+    }
+  },
+  "files": [
+    "dist/**"
+  ],
+  "scripts": {
+    "build": "tsc",
+    "check": "tsc --pretty --noEmit",
+    "clean": "rm -rf .turbo && rm -rf node_modules && rm -rf dist",
+    "dev": "tsc -w",
+    "lint": "TIMING=1 eslint src --fix",
+    "prepublishOnly": "pnpm build",
+    "test": "jest --passWithNoTests"
+  },
+  "jest": {
+    "preset": "@sushiswap/jest-config/node"
+  },
+  "dependencies": {
+    "@ethersproject/abi": "^5.7.0",
+    "@ethersproject/address": "^5.7.0",
+    "@ethersproject/bignumber": "^5.7.0",
+    "@ethersproject/solidity": "^5.7.0",
+    "@sushiswap/amm": "workspace:*",
+    "@sushiswap/math": "workspace:*",
+    "@sushiswap/chain": "workspace:*",
+    "@sushiswap/currency": "workspace:*",
+    "@sushiswap/tines": "workspace:*",
+    "ethers": "^5.7.0",
+    "ethereum-multicall": "2.15.0",
+    "tiny-invariant": "^1.2.0",
+    "tiny-warning": "^1.0.3"
+  },
+  "devDependencies": {
+    "@sushiswap/eslint-config": "workspace:*",
+    "@sushiswap/jest-config": "workspace:*",
+    "@sushiswap/prettier-config": "workspace:*",
+    "@sushiswap/typescript-config": "workspace:*",
+    "@types/node": "^17.0.33",
+    "eslint": "8.29.0",
+    "tsup": "6.5.0",
+    "typescript": "4.9.4"
+  }
+}
