+++ packages/router/src/DataFetcher.ts
@@ -3,12 +3,12 @@ import { Native, Token, Type, WNATIVE } from '@sushiswap/currency'
 import type { ethers } from 'ethers'
 
 import { Limited } from './Limited'
-import { LiquidityProviderMC, LiquidityProviders } from './liquidityProviders/LiquidityProviderMC'
-import { NativeWrapProvider } from './liquidityProviders/NativeWrapProvider'
-import { QuickSwapProviderMC } from './liquidityProviders/QuickSwapMC'
-import { SushiProviderMC } from './liquidityProviders/SushiMC'
-import { TridentProviderMC } from './liquidityProviders/TridentMC'
-import { UniSwapV2ProviderMC } from './liquidityProviders/UniswapV2MC'
+import { LiquidityProviderMC, LiquidityProviders } from './liquidity-providers/LiquidityProviderMC'
+import { NativeWrapProvider } from './liquidity-providers/NativeWrapProvider'
+import { QuickSwapProviderMC } from './liquidity-providers/QuickSwapMC'
+import { SushiProviderMC } from './liquidity-providers/SushiMC'
+import { TridentProviderMC } from './liquidity-providers/TridentMC'
+import { UniSwapV2ProviderMC } from './liquidity-providers/UniswapV2MC'
 import { MultiCallProvider } from './MulticallProvider'
 import type { PoolCode } from './pools/PoolCode'
