+++ packages/router/src/RouteCreator.ts
@@ -5,11 +5,11 @@ import { findMultiRouteExactIn, MultiRoute, NetworkInfo, RouteStatus, RToken } f
 import type { BigNumber, ethers } from 'ethers'
 
 import { Limited } from './Limited'
-import type { LiquidityProvider } from './liquidityProviders/LiquidityProvider'
-import { QuickSwapProvider } from './liquidityProviders/QuickSwap'
-import { SushiProvider } from './liquidityProviders/Sushi'
-import { convertTokenToBento, getBentoChainId, TridentProvider } from './liquidityProviders/Trident'
-import { UniswapProvider } from './liquidityProviders/UniswapV2'
+import type { LiquidityProvider } from './liquidity-providers/LiquidityProvider'
+import { QuickSwapProvider } from './liquidity-providers/QuickSwap'
+import { SushiProvider } from './liquidity-providers/Sushi'
+import { convertTokenToBento, getBentoChainId, TridentProvider } from './liquidity-providers/Trident'
+import { UniswapProvider } from './liquidity-providers/UniswapV2'
 import type { PoolCode } from './pools/PoolCode'
 import { getRouteProcessorCode } from './TinesToRouteProcessor'
