+++ packages/router/src/liquidity-providers/Sushi.ts
@@ -1,27 +1,20 @@
 // @ts-nocheck
 
 import { keccak256, pack } from '@ethersproject/solidity'
-import { ChainId } from '@sushiswap/chain'
+import { FACTORY_ADDRESS, INIT_CODE_HASH } from '@sushiswap/amm'
+import type { ChainId } from '@sushiswap/chain'
 import { ADDITIONAL_BASES, BASES_TO_CHECK_TRADES_AGAINST, Token } from '@sushiswap/currency'
 import { ConstantProductRPool, RToken } from '@sushiswap/tines'
 import { BigNumber, ethers } from 'ethers'
 import { getCreate2Address } from 'ethers/lib/utils'
 
-import { SushiPoolABI } from '../../ABI/SushiPool'
+import { SushiPoolABI } from '../ABI/SushiPool'
 import type { Limited } from '../Limited'
 import { ConstantProductPoolCode } from '../pools/ConstantProductPool'
 import type { PoolCode } from '../pools/PoolCode'
 import { LiquidityProvider } from './LiquidityProvider'
 
-const QUICKSWAP_FACTORY: Record<string | number, string> = {
-  [ChainId.POLYGON]: '0x5757371414417b8C6CAad45bAeF941aBc7d3Ab32',
-}
-
-const QUICKSWAP_INIT_CODE_HASH: Record<string | number, string> = {
-  [ChainId.POLYGON]: '0x96e8ac4277198ff8b6f785478aa9a39f403cb768dd02cbee326c3e7da348845f',
-}
-
-export class QuickSwapProvider extends LiquidityProvider {
+export class SushiProvider extends LiquidityProvider {
   poolCodes: PoolCode[]
   lastPoolCodeNumber: number
 
@@ -32,12 +25,12 @@ export class QuickSwapProvider extends LiquidityProvider {
   }
 
   getPoolProviderName(): string {
-    return 'Quickswap'
+    return 'Sushiswap'
   }
 
   async getPools(t0: Token, t1: Token): Promise<PoolCode[]> {
-    if (QUICKSWAP_FACTORY[this.chainId] === undefined) {
-      // No pools for this network
+    if (FACTORY_ADDRESS[this.chainId] === undefined) {
+      // No sushiswap for this network
       return []
     }
     const tokens = this._getAllRouteTokens(t0, t1)
@@ -48,9 +41,9 @@ export class QuickSwapProvider extends LiquidityProvider {
   _getPoolAddress(t1: Token, t2: Token): string {
     const [token0, token1] = t1.address.toLowerCase() < t2.address.toLowerCase() ? [t1, t2] : [t2, t1]
     return getCreate2Address(
-      QUICKSWAP_FACTORY[this.chainId],
+      FACTORY_ADDRESS[this.chainId],
       keccak256(['bytes'], [pack(['address', 'address'], [token0.address, token1.address])]),
-      QUICKSWAP_INIT_CODE_HASH[this.chainId]
+      INIT_CODE_HASH[this.chainId]
     )
   }
