+++ packages/router/src/liquidity-providers/Trident.ts
@@ -5,10 +5,10 @@ import { ADDITIONAL_BASES, BASES_TO_CHECK_TRADES_AGAINST, Token } from '@sushisw
 import { BridgeBento, ConstantProductRPool, RToken } from '@sushiswap/tines'
 import { BigNumber, Contract, ethers } from 'ethers'
 
-import { BentoBoxABI } from '../../ABI/BentoBoxABI'
-import { ConstantProductPoolABI } from '../../ABI/ConstantProductPoolABI'
-import { ConstantProductPoolFactoryABI } from '../../ABI/ConstantProductPoolFactoryABI'
-import { ERC20ABI } from '../../ABI/ERC20'
+import { BentoBoxABI } from '../ABI/BentoBoxABI'
+import { ConstantProductPoolABI } from '../ABI/ConstantProductPoolABI'
+import { ConstantProductPoolFactoryABI } from '../ABI/ConstantProductPoolFactoryABI'
+import { ERC20ABI } from '../ABI/ERC20'
 import type { Limited } from '../Limited'
 import { BentoBridgePoolCode } from '../pools/BentoBridge'
 import { BentoConstantProductPoolCode } from '../pools/BentoconstantProductPool'
