+++ packages/router/tsconfig.json
@@ -1,10 +1,8 @@
 {
   "extends": "@sushiswap/typescript-config/node16.json",
-  "include": ["."],
-  "exclude": ["node_modules", "dist"],
+  "include": ["src"],
   "compilerOptions": {
     "outDir": "./dist",
-    "rootDir": ".",
-    "resolveJsonModule": true
+    "rootDir": "./src"
   }
 }
